Reorganize implementation
